out with xml, in with regex; less use of `MODEL_DICT` in favor of a runtime enum
