+++ fcp/src/containment/__init__.py
@@ -7,16 +7,19 @@ from containment.structures import (
     Specification,
     VerificationSuccess,
     VerificationFailure,
-    ProofMethod,
 )
+from containment.structures.enums.data import ModelName
+from containment.structures.enums import ProofMethod
 from containment.mcp.server import mcp
 from containment.mcp.clients.experts.imp import ImpExpert
 from containment.mcp.clients.experts.proof.loop import ProofExpert as LoopProofExpert
 from containment.protocol import boundary
-from containment.fsio.experiment import run_experiments, MODEL_DICT, INCLUDE_MODELS
+from containment.fsio.experiment import run_experiments
+from containment.fsio.data import INCLUDE_MODELS
 from containment.fsio.logs import logs
 
-sonnet = MODEL_DICT["snt4"]
+sonnet = getattr(ModelName, "snt4")
+haiku = getattr(ModelName, "hku35")
 
 
 def mcp_server_run() -> None:
@@ -42,18 +45,21 @@ def test() -> None:
     async def synthesize_and_prove(
         precondition: str,
         postcondition: str,
-        model: str = sonnet.litellm_id,
+        model: ModelName = sonnet,
         max_iterations: int = 25,
     ) -> None:
         """
         Take a precondition and postcondition and ask the LLM to fill in the hoare triple and prove it.
         """
         spec = Specification(precondition=precondition, postcondition=postcondition)
-        expert = await ImpExpert.connect_and_run(model, spec)
+        expert = await ImpExpert.connect_and_run(model.value, spec)
         if expert is None or expert.triple is None:
             raise ValueError("No program found. XML parse error probably.")
         prover_pos = await LoopProofExpert.connect_and_run(
-            model, expert.triple, polarity=Polarity.POS, max_iterations=max_iterations
+            model.value,
+            expert.triple,
+            polarity=Polarity.POS,
+            max_iterations=max_iterations,
         )
         msg = ""
         if isinstance(prover_pos.verification_result, VerificationSuccess):
@@ -71,14 +77,14 @@ def test() -> None:
     async def imp_complete(
         precondition: str,
         postcondition: str,
-        model: str = sonnet.litellm_id,
+        model: ModelName = sonnet,
     ) -> None:
         """
         Take a precondition and postcondition and ask the LLM to fill in the hoare triple with an imp program
         """
         spec = Specification(precondition=precondition, postcondition=postcondition)
 
-        expert = await ImpExpert.connect_and_run(model, spec)
+        expert = await ImpExpert.connect_and_run(model.value, spec)
         print(expert.triple)
         return None
 
@@ -96,7 +102,7 @@ def contain() -> None:
         precondition: str,
         postcondition: str,
         metavariables: str = "",  # space-separated lean identifiers
-        model: str = sonnet.human_name,
+        model: ModelName = sonnet,
         proof_method: ProofMethod = ProofMethod.LOOP,
         proof_loop_budget: int = 10,
         attempt_budget: int = 5,
@@ -104,7 +110,7 @@ def contain() -> None:
         """
         Run the containment protocol at the given precondition-postcondition pair.
         """
-        model_id = MODEL_DICT[model].litellm_id
+        model_id = model.value
         specification = Specification(
             precondition=precondition,
             postcondition=postcondition,
@@ -137,18 +143,18 @@ def contain() -> None:
     async def experiments(
         proof_loop_budget: int = 50,
         attempt_budget: int = 10,
-        models: list[str] | None = None,
+        models: list[ModelName] | None = None,
         sequential: bool = False,
     ) -> None:
         """
         Run the containment protocol experiments from `data.toml`
         """
         if models is None:
-            models = INCLUDE_MODELS
+            models = [getattr(ModelName, human_name) for human_name in INCLUDE_MODELS]
         results = await run_experiments(
             proof_loop_budget,
             attempt_budget,
-            include_models=models,
+            include_models=[model.value for model in models],
             sequential=sequential,
         )
         print(results)
