+++ fcp/src/containment/structures/enums/data.py
@@ -1,0 +1,10 @@
+from enum import Enum
+from containment.fsio.data import MODEL_DICT
+
+
+class ModelName(Enum):
+    pass
+
+
+for name, model in MODEL_DICT.items():
+    setattr(ModelName, name, model.litellm_id)
